Broken build.
This commit reflects major changes in the way Deskcorder is going to work.
From now on it will be based on screen events.  Ali and I had a discussion
about it in his office just about an hour ago, and came up with this general
scheme for how things should work.

Enjoy.